ajuste de arquitetura em comments
