+++ src/likes/likes.controller.ts
@@ -8,7 +8,7 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
-import { CommentService } from '../comments/shared/comment.service';
+import { CommentService } from '../comments/service/comment.service';
 import { IsPublic } from '../autentications/decorators/is-public-decorator';
 import { Roles } from '../autentications/decorators/role-decorator';
 import { RolesGuard } from '../autentications/guards/role.guard';
